refactor: modularize interactions and fix tooltip placement
